[FEATURE] Structure for SuperPostmanSessionManager tests
